bot/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
flask==2.3.3